requirements.txt
openai 
pinecone-client
pandas
django
cohere
sentence-transformers
pinecone-client[grpc]
datasets